+++ __test__/user-actions.spec.js
@@ -1,7 +1,7 @@
 const makeFakeUser = require('./fixtures/user')
 const makeUser = require('../src/domain/user')
 const makeUserDb = require('./stubs/userdb-stub')
-const makeUserActions = require('../src/services/user-actions')
+const makeUserActions = require('../src/domain/services/user-actions')
 
 describe('user actions', () => {
   let userActions
